pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fabric-ci-test"
version = "0.1.0"
description = "Power BI visual regression testing CLI for Microsoft Fabric CI/CD"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [{name = "Filippos A. Dorezi", email = "filippos.dorezi@gmail.com"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0",
    "playwright>=1.40",
    "pytest>=7.0",
    "pytest-html>=4.0",
    "pytest-playwright>=0.7",
    "pytest-xdist>=3.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests>=2.28",
]

[project.scripts]
fabric-ci-test = "fabric_ci_test.cli:cli"

[project.urls]
Homepage = "https://github.com/FilippDor/power-bi-visual-testing-python"
Repository = "https://github.com/FilippDor/power-bi-visual-testing-python"
Issues = "https://github.com/FilippDor/power-bi-visual-testing-python/issues"
Changelog = "https://github.com/FilippDor/power-bi-visual-testing-python/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["fabric_ci_test*", "helper_functions*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/test_pbi_visual_validation.py" = ["E402"]
"tests/test_preflight.py" = ["E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
